Array impl

+++ include/baba/alias/alias.hpp
@@ -2,13 +2,14 @@
 #define BABA_ALIAS_ALIAS
 
 #include <concepts>
+#include <cstdint>
 
 namespace baba::alias
 {
 
 // unsigned integer number
 using u8 = uint8_t;
-using u16 = uint64_t;
+using u16 = uint16_t;
 using u32 = uint32_t;
 using u64 = uint64_t;
 // signed integer number
